0_aoc24_refactor/requirements.txt
prefect
anthropic
python-dotenv
requests
beautifulsoup4